main.go helpers created
