+++ cmd/cx/constant.go
@@ -1,0 +1,3 @@
+package main
+
+const VERSION = "0.8.0"
